tests/wpt/tests/performance-timeline/timing-entries-iframe.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>testing timing entries for iframe documents</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      function load() {
        promise_test(async function (t) {
          let resources = performance.getEntriesByType("resource");
          let has_resource = false;
          for ({ initiatorType, name } of resources) {
            if (initiatorType == "iframe") {
              has_resource = true;
            }
          }
          assert_true(has_resource, "iframe resource entry found");
        }, "Testing timing entries for iframe documents");
      }
    </script>
    <iframe onload="load()" src="t.html"></iframe>
  </body>
</html>
